requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.23.5
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
PyAudio==0.2.14
pyperclip==1.9.0
regex==2024.5.15
requests==2.32.3
sympy==1.12.1
tbb==2021.13.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
